Refactor modules for better organization.
