+++ crates/indexer/src/bits/auction.rs
