+++ crates/indexer/src/bits/auction_cache.rs
