+++ crates/indexer/src/bits/bidder_metadata.rs
@@ -1,0 +1,5 @@
+use crate::{prelude::*, Client, ThreadPoolHandle};
+
+pub fn get(client: &Client, handle: ThreadPoolHandle) -> Result<()> {
+    Ok(())
+}
