+++ crates/indexer/src/bits/metadata.rs
