+++ crates/indexer/src/bits/store_owner.rs
